refactor : rename doa and add export schema to false
