+++ localDatasource/src/main/java/com/example/localdatasource/roomDataBase/datasource/RecentSearchDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.example.localdatasource.roomDatabase
+package com.example.localdatasource.roomDatabase.datasource
 
 import com.example.localdatasource.roomDatabase.daos.RecentSearchDao
 import com.example.repository.datasource.local.LocalRecentSearchDataSource
